# services/backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Fitness Planner API
  summary: Cloud API for synchronized personal training plans and workout records
  description: Server-authoritative exercise catalog and immutable training plan versions, with offline-friendly idempotent
    workout synchronization.
  version: 1.0.0
paths:
...
  /api/v1/plans/current:
    get:
      tags:
      - plans
      summary: Current Plan
      operationId: current_plan_api_v1_plans_current_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanVersionOut'
      security:
      - HTTPBearer: []
  /api/v1/plans/{plan_version_id}:
    get:
      tags:
      - plans
      summary: Plan Detail
      operationId: plan_detail_api_v1_plans__plan_version_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_version_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanVersionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/exercises:
    get:
      tags:
      - catalog
      summary: List Exercises
      operationId: list_exercises_api_v1_exercises_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExercisePage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/equipment:
    get:
      tags:
      - catalog
      summary: List Equipment
      operationId: list_equipment_api_v1_equipment_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/readiness:
    post:
      tags:
      - readiness
      summary: Create Readiness
      operationId: create_readiness_api_v1_readiness_post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadinessUpsert'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - readiness
      summary: List Readiness
      operationId: list_readiness_api_v1_readiness_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 30
          title: Limit
      - name: local_date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Local Date From
      - name: local_date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Local Date To
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Deleted
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EquipmentOut:
      properties:
        id:
          type: string
          title: Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        category:
          type: string
          title: Category
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - code
      - name
      - category
      - version
      - created_at
      - updated_at
      title: EquipmentOut
    EquipmentPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EquipmentOut'
          type: array
          title: Items
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
          default: false
      type: object
      title: EquipmentPage
...
    ExerciseAlternativeOut:
      properties:
        id:
          type: string
          title: Id
        exercise_id:
          type: string
          title: Exercise Id
        alternative_exercise_id:
          type: string
          title: Alternative Exercise Id
        sort_order:
          type: integer
          title: Sort Order
          default: 0
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - exercise_id
      - alternative_exercise_id
      - version
      - created_at
      - updated_at
      title: ExerciseAlternativeOut
...
    ExerciseOut:
      properties:
        id:
          type: string
          title: Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        body_part:
          type: string
          title: Body Part
          default: ''
        equipment_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Equipment Id
        default_sets:
          type: integer
          title: Default Sets
          default: 0
        rep_min:
          type: integer
          title: Rep Min
          default: 0
        rep_max:
          type: integer
          title: Rep Max
          default: 0
        rep_unit:
          type: string
          title: Rep Unit
          default: reps
        cues:
          type: string
          title: Cues
          default: ''
        common_mistakes:
          type: string
          title: Common Mistakes
          default: ''
        definition_version:
          type: integer
          title: Definition Version
          default: 1
        alternatives:
          items:
            $ref: '#/components/schemas/ExerciseAlternativeOut'
          type: array
          title: Alternatives
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - code
      - name
      - version
      - created_at
      - updated_at
      title: ExerciseOut
    ExercisePage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ExerciseOut'
          type: array
          title: Items
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
          default: false
      type: object
      title: ExercisePage
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PlanDayOut:
      properties:
        id:
          type: string
          title: Id
        plan_version_id:
          type: string
          title: Plan Version Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        sort_order:
          type: integer
          title: Sort Order
        slots:
          items:
            $ref: '#/components/schemas/PlanSlotOut'
          type: array
          title: Slots
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - plan_version_id
      - code
      - name
      - sort_order
      - version
      - created_at
      - updated_at
      title: PlanDayOut
...
    PlanSlotOptionOut:
      properties:
        id:
          type: string
          title: Id
        plan_slot_id:
          type: string
          title: Plan Slot Id
        exercise_id:
          type: string
          title: Exercise Id
        equipment_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Equipment Id
        is_preferred:
          type: boolean
          title: Is Preferred
        sort_order:
          type: integer
          title: Sort Order
        set_count:
          type: integer
          title: Set Count
        rep_min:
          type: integer
          title: Rep Min
          default: 0
        rep_max:
          type: integer
          title: Rep Max
          default: 0
        rep_unit:
          type: string
          title: Rep Unit
          default: reps
        duration_seconds_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Seconds Min
        duration_seconds_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Seconds Max
        rir_min:
          anyOf:
          - type: number
          - type: 'null'
          title: Rir Min
        rir_max:
          anyOf:
          - type: number
          - type: 'null'
          title: Rir Max
        is_per_side:
          type: boolean
          title: Is Per Side
          default: false
        prescription_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Prescription Text
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - plan_slot_id
      - exercise_id
      - is_preferred
      - sort_order
      - set_count
      - version
      - created_at
      - updated_at
      title: PlanSlotOptionOut
    PlanSlotOut:
      properties:
        id:
          type: string
          title: Id
        plan_day_id:
          type: string
          title: Plan Day Id
        position:
          type: integer
          title: Position
        body_part:
          type: string
          title: Body Part
        cues:
          type: string
          title: Cues
          default: ''
        options:
          items:
            $ref: '#/components/schemas/PlanSlotOptionOut'
          type: array
          title: Options
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - plan_day_id
      - position
      - body_part
      - version
      - created_at
      - updated_at
      title: PlanSlotOut
...
    PlanVersionOut:
      properties:
        id:
          type: string
          title: Id
        plan_id:
          type: string
          title: Plan Id
        plan_name:
          type: string
          title: Plan Name
        version_number:
          type: integer
          title: Version Number
        status:
          type: string
          title: Status
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
        snapshot_json:
          anyOf:
          - type: string
          - type: 'null'
          title: Snapshot Json
        weekly_frequency:
          type: integer
          title: Weekly Frequency
        min_rest_days:
          type: integer
          title: Min Rest Days
        fatigue_threshold:
          type: integer
          title: Fatigue Threshold
        initial_reduced_weeks:
          type: integer
          title: Initial Reduced Weeks
        initial_set_count:
          type: integer
          title: Initial Set Count
        rules:
          additionalProperties: true
          type: object
          title: Rules
        days:
          items:
            $ref: '#/components/schemas/PlanDayOut'
          type: array
          title: Days
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - plan_id
      - plan_name
      - version_number
      - status
      - weekly_frequency
      - min_rest_days
      - fatigue_threshold
      - initial_reduced_weeks
      - initial_set_count
      - version
      - created_at
      - updated_at
      title: PlanVersionOut
...
    ReadinessOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        local_date:
          type: string
          format: date
          title: Local Date
        fatigue_score:
          type: integer
          title: Fatigue Score
        sleep_quality:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sleep Quality
        pain_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Pain Notes
        soreness:
          anyOf:
          - type: integer
          - type: 'null'
          title: Soreness
        stress:
          anyOf:
          - type: integer
          - type: 'null'
          title: Stress
        motivation:
          anyOf:
          - type: integer
          - type: 'null'
          title: Motivation
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        metrics:
          additionalProperties: true
          type: object
          title: Metrics
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - user_id
      - local_date
      - fatigue_score
      - version
      - created_at
      - updated_at
      title: ReadinessOut
    ReadinessPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ReadinessOut'
          type: array
          title: Items
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
          default: false
      type: object
      title: ReadinessPage
    ReadinessUpsert:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        local_date:
          type: string
          format: date
          title: Local Date
        fatigue_score:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Fatigue Score
        sleep_quality:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Sleep Quality
        pain_notes:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Pain Notes
        soreness:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Soreness
        stress:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Stress
        motivation:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Motivation
        notes:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Notes
        metrics:
          additionalProperties: true
          type: object
          title: Metrics
        expected_version:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Expected Version
      type: object
      required:
      - id
      - local_date
      - fatigue_score
      title: ReadinessUpsert
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer